[Feature #164697688] Made more corrrections to develop branch
