+++ controllers/Staff/debitAccounts.js
@@ -1,7 +1,8 @@
 // @ts-ignore
 import logic from '../index';
 import data from '../../db';
+import services from '../../services';
 
-logic.debitAccount = () => {}
+logic.debitAccount = () => {};
 
 export default logic.debitAccount;
