+++ controllers/User/createBankAccount.js
@@ -1,6 +1,7 @@
 // @ts-ignore
 import logic from '../index';
 import data from '../../db';
+import services from '../../services';
 
 logic.createBankAccount = () => {};
