+++ helpers/index.js
@@ -1,14 +1,4 @@
-class Helper {
-  async validateEmail(email) {
-    this.emailPattern = /^(([^<>()[\].,;:\s@"]+(\.[^<>()[\].,;:\s@"]+)*)|(".+"))@(([^<>()[\].,;:\s@"]+\.)+[^<>()[\].,;:\s@"]{2,})$/i;
-    return this.emailPattern.test(email);
-  }
-
-  async validatePassword(password) {
-    this.passwordPattern = /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]{8,}$/;
-    return this.passwordPattern.test(password);
-  }
-}
+class Helper {}
 
 const helper = new Helper();
 export default helper;
